+++ backend/app/api/routes/whatsapp.py
@@ -128,7 +128,7 @@ def _format_alertar(conn):
     from app.services.monitoring_service import run_full_check
     from app.models.base import Configuracion
     alert_umbral_conf = Configuracion.query.filter_by(clave='alert_umbral').first()
-    alert_umbral = int(alert_umbral_conf.valor) if alert_umbral_conf else 30
+    alert_umbral = float(alert_umbral_conf.valor) if alert_umbral_conf else 30.0
     result = run_full_check(conn, alert_umbral)
     if result.get('status') != 'success':
         return f"❌ Error ejecutando chequeo: {result.get('message')}"
